On Mac OS X, simulate MSX Insert key using Cmd+I
Apple keyboards don't have an Insert key.
It seems that all single keys that exist on all international keyboard
variants are already taken, so use a key combo to produce Insert.

The combo only works if Cmd is pressed first and is released last.
Since this is the typical way to type such a combo and making the
code robust against other press/release orders would make it much
more complex, I think this is acceptable.