More robust check for empty directory
The check for
   do not show pool shortcuts when that pool is empty
did literally check for an empty sub-directory. Though in case of the default
system pool directory there is a README file in that directory. When actually
showing the content of the directories, the files are still filtered on
extension. So the effect was that in the menus we still showed an effectively
empty pool.

Fixed by applying the same filter to the is-empty check.