Simplify parsing of softwaredb.xml <remark> tag
An entry in softwaredb.xml looks like this (simplified):
   <software>
      ...
      <remark>...</remark>
      <dump>
         ...
         <remark>...</remark>
      </dump>
      <dump>
         ...
         <remark>...</remark>
      </dump>
   </software>
So a <remark> tag can appear in two different locations. On the level of
'software' (meaning it applies to all dumps of this software) or at the level
of a specific 'dump'. Before this patch, the parser would for each dump merge
the remarks at both levels and store the combination. The parser would also
merge multiple <remark> tags at the same level.

Though in almost 9 years that we use this softwaredb.xml format, we never ever
had a <remark> tag at the 'software' level. We also never had multiple <remark>
tags at the same level. So this allows to simplify the parser.

This may seem like a trivial simplification (hardly worth doing). However,
if we never need to merge multiple remark strings into one, we can store the
remark using a string_ref object instead of a std::string. And this resulted
in the following differences:
 - Use about 50kB less heap memory.
 - Perform about 2600 less heap allocations during startup.
 - Parsing got about 10% faster, parsing softwaredb.xml takes about 10% of
   total startup time. So this results in roughly 1% faster startup.
And of course it simplifies the code.