Replace chain of if-statements with switch-statment
While parsing softwaredb.xml, a lot of time was spend doing string comparisons.
Most of those came from a if-chain like like:
          if (tag == "system") { ..
   } else if (tag == "title") { ...
   } else if (tag == "year") { ...
   ...
So often we had to do a whole lot of comparisons before we found the matching
block. Now we instead switch on the first character of the string and then only
do one (occasionally two) string comparison.

Speeds up total startup time by a slightly less than 0.5% (not much but still
significant for such a small change).