Simplify softwaredb.xml
Remove xml:lang="en" attribute (it's present on the <title> attribute). It
currently never has a different value than "en", and openMSX doesn't use this
attribute at all.

The openMSX parser accepts both:
  <remark><text>blabla</text></remark>
  <remark>blabla</remark>
I think the original idea for the first format was to allow multiple
translations of the same remark. Though in 9 years we never made use of this,
and we don't have plans to do so in the future. This patch removes the <text>
tag.

This only makes a small difference in startup speed (about 0.3% faster). It
also reduces the size of the softwaredb.xml by about 34kB, this might be
important on embedded systems because this file is kept in memory (and modified
inplace) for the whole openMSX session.

I've already coordinated with Vampier to update his softwaredb.xml creation
scripts.