Don't lazily create 'reversebar_fadeout_time' setting
Before this patch, in TAS mode (set mode TAS) when you start openmsx, you
triggered the following problem:
	warning: Error executing delayed command: can't read
	"::reversebar_fadeout_time": no such variable
The problem was that "reversebar_fadeout_time" was created in "_reverse.tcl",
but that script is only loaded on demand (and there is no check for settings
yet), so the setting didn't exist yet when the mode.tcl script tried to use it.

Fixed by moving the creation of "reversebar_fadeout_time" to "reverse.tcl".
(Luckily) we already had a non-lazy part of the reverse script.