Use c++11 way of generating gaussian random numbers
I replaced our own Math::gaussian2() routine with the (new in c++11)
std::normal_distribution stuff. Main advantage is that the new code is both
shorter and more descriptive. We now also perform the calculation using float
instead of double (should still be plenty).

These random numbers are only used to initialize some data structures, so
performance of calculating them is not that important. The next paragraph is
just for fun ;-)

I compared the performace of the following 4 variants in a standalone program:
(times normalized to a, lower is better)
  a) gaussian2, double            : 100  <- before patch
  b) gaussian2, float             :  81
  c) normal_distribution, double  : 104
  d) normal_distribution, float   :  69  <- after  patch