Replace std::deque with cb_queue in some places
Originally we used a std::deque because we both need push_back() and
pop_front(). Now that we have a (dynamic) circular buffer it's more efficient
to use that data structure. Though none of these routines are critical for
peformance.

I expected to see a reduction in code size, because most operations on cb_queue
are simpler than those on std::deque. And if you look at the generated code
this is indeed the case. Though because the routines are smaller, gcc decides
to inline them more often. The netto effect is a small increase in code size.
(We could annotate the code to make the inliner less aggressive, but that's
probably not worth the effort).