Replace data structures in OggReader
Where possible replace deque/list with vector or cb_queue. I kept one instance
of std::list because it also used erase(). It's likely possible to rewrite this
so that we only remove elements at the front or back. But that's something for
a different patch (also low priority).

Saves about 4kB code size.