Ran clang-modernize on our codebase and applied (some of) the suggestions
The most common changes are:
- use range-based for-loop also for arrays
- use auto in a few more places
- pass sink-arguments by value (instead of const-reference) to (better) take
  advantage of move-semantics

Clang-modernize also made a _whole_lot_ of suggestions where the new c++11
'override' keyword can be used. I didn't apply those (yet) because I'm not
convinced yet it's really useful and it are _a_lot_ of changes that are anyway
better done in a separate patch.