Added xxhash (non-cryptographic) hashing function
See here for more information on xxhash:
  http://code.google.com/p/xxhash/
Summary: xxhash is a non-cryptographic hashing function. It produces high
quality hash values at very fast speeds.

Use this new hash function in the implementation of StringMap. Before we
used the 'FNV' hash function there.

Compared to 'FNV', 'xxhash' can be 10x faster. Xxhash achieves this speed by
working at words (32-bits) or even 4 words at a time instead of byte-per-byte.
Of course this is only possible on large strings. In openMSX we typically hash
small strings, because of this I only measured a speedup of about 33%.

This patch also makes a small step of allowing StringMap to hash/compare
strings in a case-insensitive way.