Speedup string->RomType conversion
(To me surprisingly) 10% of the softwaredb.xml parsing time was spend in
string->RomType conversion. This patch speeds up that routine by more than a
factor 3x, resulting in about 1% faster startup speed.

Initially we used a simple 'std::map<std::string, RomType>' to perform this
conversion. Though this isn't a very cache-friendly data-structure. Later we
changed this to binary-search in a sorted vector. This performed better, but
it still requires O(log(N)) string comparisons, this is relatively expensive
(also because we need case-insensitive comparisons). This patch performs the
lookup using a hash-table (StringMap class). This typically only requires O(1)
comparisons.