Revert "Use variadic templates in our version of make_unique()"
This reverts commit 797118da98e593b4cb39ff26c91f533ab38f5d0a.

Although this worked perfectly in gcc, with visual studio it caused lots of
compiler errors. I'm not 100% sure, but it seems the vs errors are not directly
related to the variadic templates but instead there seems to be a lookup clash
with std::make_unique() that's already shipped with vs.

Maybe I'll think of a better solution in the future. For now i'll simply revert
this change.