Fix Tcl error when changing horizontal stretch in OSD menu
The code to refresh the menu was overlooked when in commit fd009f09 the
video menu contents were made dynamic (skipping irrelevant settings).

There is a proc specifically to refresh the menu; use that instead of
the close-then-reopen trick. It is not only simpler but it also
preserves the cursor position.