More efficient XMLElement::addChild()
XMLElement is our class that represents nodes in an XML tree. One method on
this class is 'XMLElement& addChild(XMLElement child)' this adds (a copy) of
one XMLElement as a child node of another XMLElement. Note that we use value-
instead of referencer-semantics. So child nodes are not references (or
pointers) to other nodes. Instead child nodes are stored (by value) inside the
parent object itself. (See revision 884923d74 for more details. It also
explains that usually copies are optimized to moves).

Very often we use the addChild() method like this:
	myNode.addChild(XMLElement("name", "data"));
So this will:
 - Create a temporary XMLElement object.
 - Insert a copy of that object into myNode (possibly this copy is optimized to
   a move).
 - Destroy the temporary object.
It would be more efficient if we directly construct the new XMLElement in the
correct position. And that's what this patch allows:
	myNode.addChild("name", "data");
Internally this uses emplace_back() to really construct the object in the
correct (final) position. See also patch bb01231.

If you look at small pieces of code there can be big differences. E.g.
generated code for createXML() in CassettePlayer.cc shrinks from ~1000 to ~100
bytes. But in other cases this smaller code is offset by more inlining, so
overal this patch is code-size neutral. The code should be faster, but because
XMLElement is not used that much anymore this only has a small impact on the
total speed. This also reduced the amount of heap allocations during startup
by about 50.