Leave 'SUSHOLD' envelope state when EGTYPE changes
The YM2413 ADSR (attack-decay-sustain-release) envelope generators can work in
two main modes controlled by the 'EGTYPE' bit (bit 5 in registers 0 and 1). The
main difference is whether there is a sustained or a declining output level
when 'key' remains =1 (see YM2413 documentation for a much better description).

If the output level is kept constant, the 'Slot' object is in the state
'SUSHOLD' and it stays in that state until key=0. Though when instrument
parameters are changed (the 'Patch' object in the code) it's possible that also
the EGTYPE changes. In that case we must leave the 'SUSHOLD' state.

There was already code to detect this condition, but only on the code path that
explicitly changes the instrument setting registers (writes to register 0 and
1). It's also possible to implicitly change the instrument settings by changing
the rhythm mode (register 14). And in that code path this check was missing.

Side note: on real hardware the whole envelope state is likely continuously
being evaluated (or at least once each sample is calculated), not only on
specific register writes. Though in software emulation it's much cheaper to
only check specific conditions when they potentially could change. But then
of course you have to make sure to cover all possible cases.

This patch fixes bug '#511 Hihats sounding while they shouldn't in BPS Tetris'.