Fixed (wrong) warning when sha1sum were not using lowercase
The ROM loading code uses several ways to locate the desired rom on the users
filesystem (try several sha1 values in FilePool, or try several filenames,
possibly derived from savestate information, see Rom::init() for details). When
a ROM is found we still verify whether it has a sha1 that's equal to one of the
listed sha1 values in the config file. If it doesn't match we print a warning.

Internally in the code we have two representations for sha1 values:
 - a compact 160-bit value (stored as 5x32-bit integers)
 - an unprocessed string

Before this patch the verify-matching-sha1 code of the 1st paragraph would take
the sha1 of the found rom, convert that to string and compare that to all the
sha1 strings of the rom config. Though it did a normal (case-sensitive) string
comparison (and that's a bug). This patch instead converts all strings to sha1
objects and compares those. This fixes the case-comparison bug, but has the
additional advantage that it also catches invalid sha1 values in the config
file (not exactly 40 hex-digits).