Fixed MIDI output on OS X
Previously, openMSX was sending the MIDI data one byte at a time.
However, CoreMIDI wants to receive full messages (except for SysEx
which can be sent in chunks). For some reason, many applications
received the MIDI output from openMSX correctly nevertheless, but
a few applications had problems.

Thanks to Laurens Holst for finding and confirming the cause of the
incompatibilities.