Fix crash when using wav files on a visual studio build
OpenMSX build using visual studio (including version 2013) would crash when
using .wav files (e.g. in the cassetteplayer). Strictly speaking our code was
fine, but because of a limitation in visual studio it generated wrong code.
This patch adds a workaround.

More in detail: revision 4b50e71c made 'WavData' objects move-only (so not
copyable). Though because vs2013 is not fully c++11 compliant yet, these
objects were still copyable (and not movable) in visual studio. But then both
the original and the copied object assume they 'own' the same memory and this
causes a too early free (and eventual double-free) of that memory.

This patch also adds a tweak in the MemBuffer code so that future similar
problems will result in a (visual studio) compile-error instead of silently
generating wrong code.