Fix(?) crash on windows during video recording
On a 32-bit visual studio windows build, when doing a 320-pixels wide video
recording of a 512-pixels wide msx screen mode, you could get a crash. This was
caused by an unaligned SSE2 memory write. Because the memory returned by
malloc() was only 8-bytes aligned (and SSE2 requires 16 bytes).

On 64-bit (windows and linux) systems the memory returned by malloc() is always
16-bytes aligned. On 32-bit gcc builds, SSE2 is (by default) disabled. This
explains why this crash was so hard to reproduce.

Fixed by extending the MemBuffer class with an (optional) parameter that
specifies the memory alignment requirement. See comments in that code for more
details.

An alternative fix could have been to not extend MemBuffer but instead directly
use our MemoryOps::mallocAligned() routines from the ZMBVEncoder code. I choose
for the current approach because it may allow to in the future also replace the
existing mallocAligned() users with the more high-level MemBuffer approach.
And this is a better (safer) c++ style of resource handling.