Cache TigerTreeHash calculations
In the previous openMSX release we stored the sha1sum of harddisk images in a
savestate (and thus also in a reverse snapshot). This had the disadvantage that
each time you wrote to the harddisk, the whole sha1sum had to be recalculated
on the next savestate (or snapshot). So we switched sha1sum->TigerTreeHash(TTH)
because TTH can be calculated incrementally.

Each time you load a savestate (or when using reverse) we calculate the hash of
the harddisk image and compare that with the hash that's stored in the
savestate. In the previous release, with sha1, the Filepool-sha1-cache stuff
kicked in so that we don't actually needed to recalculate the full sha1sum if
the modification date of the image wasn't change. But for TTH we don't have
such a caching mechanism. Or not until this patch.

This patch keeps an in-memory cache of all TTH calculations. It's also based on
the file modification date (so if you e.g. externally change the file, the hash
is properly recalculated). Unless the FilePool stuff, the TTH cache is not
stored on disk. So if you restart openMSX the cache is empty again. For now
that's fine (TTH is used much less frequently than sha1).

The main motivation for this change was using reverse in combination with the
IDE extension. Before this patch when pressing PgUp there was a short delay
(less than a second, but still very noticeable). Now it's again almost
instantaneous (same as without IDE extension).