Fixed bug in cb_queue::checkGrow()
When we're about to insert new data in a cb_queue, but that queue is already
full, we first create new room by doubling the capacity of the queue. Though
when the initial capacity is zero, then doubling doesn't actually help much ;)

I guess I only tested this with queues that did have a non-zero initial
capacity.