Fixed potential openmsx crash (when MSX crashed)
When the MSX software crashes it sometimes happens that invalid commands are
send to the FDC (e.g. when diskrom is still selected and the z80 goes in an
infinite loop that keeps on growing the stack). And sometimes this will execute
a read-track command when no diskdrive is selected. Exactly this happened in a
development(?) version of trilo-tracker. Thanks huey for reporting this bug.

In the above scenario the method DummyDrive::readTrack() was called. But that
method was implemented using the 'UNREACHABLE' macro. Meaning that we promise
the compiler that this method will never be called, so the compiler can
implement/optimize it however it wants (so typically the compiler generates no
code at all for such methods, not even a 'ret' instruction). However if that
promise is broken, so if the method does get called, anything can happen,
including crashing the application.

When using a WD2793 controller the method can indeed never be called: the
WD2793 code has explicit checks that the disk is ready before calling
readTrack(). But in the TC8566AF code there was no such protection.

Fixed by implementing DummyDrive::readTrack(). Using UNREACHABLE here was
anyway only a very minor (and thus also wrong) optimization.