Temporary(?) disable SSE2 on 32-bit visual studio build
We have strong suspicions that the crash in ticket
  #515 Crashes on 32-bit Windows with scale factor 3 and SDL renderer
is caused by a misaligned SSE2 memory access. Fixing this type of problem
shouldn't be hard when a back-trace of the crash can be made. Unfortunately in
the last few weeks there hasn't been much progress on this front. So for now
I'm disabling SSE2 in the 32-bit visual studio build. This will at the very
least allow to confirm/deny our suspicion. And we can always revert this patch
in the future when we get more help from windows developers to really search
for the root cause.
This bug was blocking the 0.10.1 release. This patch (likely) can serve as a
workaround.