Tweak behavior of -control command line option
This is a follow-up on e05b519f4c
    Process -control command line option last
See that commit message for details. This commit triggered a regression:
  bug #522 Double output in Catapult's Output Tab for (most?) errors

One set of error messages was send via the control-connection, the other via
stderr (the latter is causing the regression). When there is no -control
connection we display error/info messages on stdout/stderr (this is done by the
StdioMessages class). If there is a connection we cannot send to stdout because
that would interfere with the control-protocol. So there is code that only
creates a StdioMessages object when there's no -control connection. Though this
check was now done too early (because processing -control was moved later). So
the StdioMessages object was always created, resulting in double messages.

I tried moving the (conditional) creation of StdioMessages later, but that
caused other problems (catapult not being able to detect the openMSX version, I
didn't look into it further). But actually you _do_ want to establish the
control connection early so that possible error messages triggered during
openMSX startup can be send over the connection. But also (like explained in
e05b519f4c) we cannot start accepting external commands too soon. To solve this
the function of CliConnections is now logically split in two. They can be
established early and send messages back to the external application, but
commands received over the connection are only accepted late in the startup
process.