Tweaks in _toggle_freq.tcl
In constructs like
   if {[expr <some-expression]} {
the 'expr' part is superfluous because the first argument of an if command is
already interpreted as an expression. So it can be simplified to
   if {<some-expression>} {

In this specific case the script could be simplified even more by using ?: and
by using the peek/poke/vdpreg helper procs.