Worked around (false positive) Coverity warning
The function call
   glGetIntegerv(GL_SCISSOR_BOX, &xo);
was used to initialize 4 successive variables: xo, yo, wo, ho
Though Coverity only assumed that this call changes xo.

This code did rely on the specific order of the xo, yo, wo, ho member
variables. I believe this code is correct (c++ standard guarantees layout of
the member vars), but it is not very robust (there even was a comment in the
code to not change the order of these 4 members).

So to silence the warning (and to make the code more robust) I transformed
those 4 variables into an array (with 4 elements).