Add assert to help coverity
Coverity (wrongly) complained that 'ps' could be used to index outside of an
array of 4 elements. Though if you look closer, at that point in the code,
'ps' must be '0 <= ps <= 3'. I've added an assert to document this.

I'm not sure Coverity is smart enough to understand this assert. But let's
see what happens. Independent of Coverity, this assert is a good idea anyway.