Initialize modeBase to avoid Coverity warning
This member was not yet initialized in the constructor but it was before first
use (although not obvious to see). But it doesn't cost much to initialize it
anyway.

[ After fixing this issue, I saw that Coverity found a lot more constructors
that don't initialize all members. I checked them all and as far as I can see
none of them are real bugs, so I didn't try to silence all those warnings.
Instead Coverity has a feature to mark an issue as 'Intentional'. So let's see
how well that feature works for future Coverity-scans. ]