Made 'workused' a local- instead of a member-variable
Coverity complained that 'workUsed' was not initialized in the ZMBVEncoder
constructor. That's because this variable is only needed for the duration of
the compressFrame() method. So it could actually be a local variable of that
method. The only disadvantage is that this requires passing an extra parameter
to a few other methods. (But if the compiler is smart enough this can be
optimized away and lead to both better and clearer code than before this
patch).