Slightly rewrite to suppress Coverity warning
There was a construct like this in the code
    &*bufferStorage.end()
This constructs a pointer to the one-past-end element of the buffer. Though
syntactically this dereferences the end-iterator and that's not allowed. In
practice, in the generated code, this dereference will not be present, so it
will work just fine.

Next to Coverity it's likely that (some implementations of) STL in debug mode
will also complain about this.

Anyway, it's easy enough to change this into legal C++ code:
    bufferStorage.data() + bufferStorage.size()