Change type of 'status' variable
Bug discovered by Coverity.
This code is ported from blueMSX, though this port is incomplete and this
version of the emulated chip doesn't do anything useful yet (maybe apart
from fooling some MSX detection routines).

Anyway the bug was that the expression
	status |= ST_INT;
doesn't do anything because 'status' is an 8-bit variable and the constant
'ST_INT' has the value 0x800.

In the original blueMSX code 'status' was a 32-bit value. Though if I quickly
look at the code only the lower 8 bits of this variable are really used. The
higher bits are written to, but they are never used again.

For now I've changed 'status' to a 32-bit value (as it was in blueMSX). But
this doesn't affect the functionality of the code. So either there is a bug or
the code can be simplified. Because this chip (YM2148) is not often used and
because it's anyway not yet functional, investigating this further is low
priority for me.

--

There is a similar issue with
	if (command & CMD_WRINT)
'command' is an 8 bit variable and the value of the constant 'CMD_WRINT' is
0x100. This is also the case in the blueMSX code. There is also no code that
ever sets the higher bits of this variable. This issue isn't fixed.