Fix(?) #524 Crash on Windows when switching scale_algorithm
Vampier created a backtrace of this crash (thanks!). It points to this line:
  VLA(unsigned, verticals, srcNumLines * srcWidth);
The strange thing is that a bit earlier in the code there is an almost
identical VLA() for 'horizontals' and that works fine. Though I *think* the
problem is the following:

The MLAAScaler::scaleImage() method allocates quite a few large buffers on the
stack. The size of these buffers is proportional to the amount of to-be-scaled
MSX pixels. In case of an high-resolution MSX image (screen0/80, screen6,
screen7) The total required buffer space is more than 1MB! And the line above
is the allocation that pushes the stack size over 1MB.

On windows the default stack size is only 1MB. On Linux it's typically 8MB. So
this may explain why it only crashes on windows and only on hi-res MSX screen
modes.