Fixed wrong error message about end-of-tape
The game "Chuckie Egg" doesn't turn off the cassette motor after the end of the
tape is reached. So you'll get a message that the end of the tape was reached.
If you use reverse around this point you could get a message that the loaded state
contains a tape position beyond the end of the tape.

This bug was triggered because on loadstate we didn't recalculate the moment in
time when the end-of-tape will be reached correctly. Fixed now.