Fixed -romtype command line option regression
This works:
	openmsx -carta myrom.rom -romtype KonamiSCC
This used to work as well, but was now broken:
	openmsx myrom.rom -romtype KonamiSCC

The regression was introduced in e05b519f4c where I (unintentionally) parse
command line filenames in a different pass than the -romtype option. Fixed now.

Actually in the future it might be a good idea to break backwards compatibility
and just parse the commandline from left to right. Lots of the historical
reasons to have multi-pass parsing are no longer valid (e.g. possibility to
specify filetypes in the settings file).