In DirAsDsk prefer using 'myfile.txt' over 'myfile.txt~'
Before this patch when the host directory contained both files
  myfile.txt  myfile.txt~
then only one of these two ended up in the virtual MSX disk image. Which one
wasn't defined (depends on the order in which the host filesystem enumerates
the files).

This patch adds some heuristics to prefer the 1st version. See comments in
code for more details.