Revert "install: Install symlink with absolute path."
This reverts commit a3e79e758a23dc9164b474999ebf2cdd223a8cec.

If the installation location is a relative path, openMSX is building
some kind of package, which is intended to be either relocatable or
installed in a fixed location outside the openMSX build tree.
Therefore, creating a symlink to an absolute path inside the openMSX
build tree is the wrong approach.