Compile against the default SDK on Mac OS X
We had problems in the past when compiling against the default SDK
because we would link to library versions not available in older Mac
OS versions. However, I think this problem occurs because we linked
to libraries in /usr/lib directly.

This commit removes the linking to libraries in /usr/lib: even if
Mac OS X has the library in the base system, we'll build our own copy
using the 3rdparty system. We do still use systemwide libraries that
are packaged as frameworks, under the assumption that those will be
backwards compatible.