Reduce size of LUT for rapidsax
The rapidsax code used 10 tables of each 256 booleans. C++ stores one boolean
in one byte. Though by using bitmasks we can actually fit 8 boolean into one
byte. And when we choose the bitmasks in a clever way, for this particular
case, it's even possible to fit all 10 original bool-tables into only one
byte-table. So reducing the memory requirement from 2560->256 bytes. And
testing whether a byte is zero or whether a particular bit is zero is just as
fast.

This change was inspired by this article:
	http://aosabook.org/en/posa/parsing-xml-at-the-speed-of-light.html
This describes the design of the pugixml parser. Benchmarks (found of the
pugixml website) show that pugixml is a tiny bit faster than rapidxml (on which
our xml parser is based). It's possible that there are more tricks in pugixml
that can improve our code. Though xml parsing is only a small fraction of the
total openmsx startup cost anymore, so optimizing it further is much less
beneficial than what it used to be. But of course feel free to do so ;-)