Suppress Coverity issue 1202612
The intention of the original code was to have no error checking in a
non-assert build. Though this construct wasn't very portable and Coverity
doesn't understand 'UNREACHABLE'. Suppressed Coverity warning by just leaving
the check in place even in an optimized build. It anyway shouldn't have a
big impact on performance.