Minor cleanups in the logo PNGs
Removed some gray dots and almost-white stripes on the penguin's
belly. Touched up rough edges on the (from the penguin's view) right
flipper and left side of the head. This was done on the 256x256 version
of the icon.

Regenerated 128x128 and 64x64 versions of the icon from 256x256, using
cubic scaling in GIMP. Manually adjusted the eyes on 64x64.

Applied PNG optimization to the logo images.