[3/13] map->vector: ResampleHQ coefficient tables
ResampleHQ shares coefficient tables between all sound devices with the same
resample ratio. It keeps those tables in map (indexed of ratio). There aren't
that many different sound chips in a MSX computer and many even share the same
resample ratio. This map typically contained only 1 or 2 (or 4 max) elements.
So this patch replaces that map with an (unsorted) vector.

Insert/erase/lookup in this vector are all O(n). For erase we use a small
tweak: instead of directly doing std::vector::erase() we swap the element
with that last element in the vector and then do a pop_back(). That doesn't
preserve the relative order of the remaining elements, but that's OK.