[7/13] map->vector: polymorphic saver registery
The serialization code maintains a map of all polymorphic class saver code.
This map is first constructed (without being queried) and later only queried
(without adding/removing items). So instead of using a map, it's cheaper to
first add items to an unsorted vector, then sort that vector and in the end
binary search that vector.