[8/13] map->vector: (un)aligned memory pair tracking
In case we can't use a system routine to allocate aligned memory (or in debug)
mode we manually keep track of corresponding aligned and unaligned memory
pointer pairs. Typically there are only very few items in this map, and
typicallythay are even inserted/removed in (approximatly) LIFO order. So using
an unsorted vector instead of a map is more efficient.