[9/13] map->vector: IPSPatch
The data structure used inside a IPSPatch was basically a map of start-address,
replacement sequence. Once that map is constructed it's no longer changed, but
queried a lot of times. So using a sorted vector instead of a map is more
efficient.

Also during construction it's required to have the structure in sorted order
(thus we can't use the strategy of build, then sort once, then query).

It's possible that the construction phase is now slower, but the part that
matters a lot more, using the structure, is certainly faster. And typical IPS
files don't contain several hundreds of regions, so it's even likely that also
construction is faster.