[11/13] map->vector: MSXMixer SoundDeviceInfo
MSXMixer maintained a mapping from each registered SoundDevice to some extra
information about that device. That structure is updated rarely (when new
devices appear or disappear), but we iterate many times over it. So it's
cheaper to store this in a vector than in a map. Actually the only time we
use O(log(n)) lookup is when we unregister a device, so we can just as well
use an unsorted vector. (And of course also this structure doesn't contain
that many elements).