[12/13] map->vector: android pressed keys
For the Android hackers keyboard we need to remember which keys are pressed but
haven't been released yet (see comments in code for details). Typically only
very few keys are pressed at the same time (and in the most extreme case only
~100). So it's better to store this is a (unsorted) vector than in a map.

Note: I could test-compile this patch, but couldn't actually run it.