[13/13] map->vector: breakpoints
The CPU emulation code keeps a (multi)map of breakpoints indexed on PC.
Because this structure is much more frequently queried than changed it's
better to use a (sorted) vector.

This patch is slightly different than the previous ones because this is the
one only that replaces a multi_map (instead of a map).

I considered using an unsorted vector with linear instead of binary search
because there rarely will be more than a few dozen breakpoints (and in that
case linear search is faster). But, who knows, someone might have a good reason
to set a thousand or more breakpoints. And anyway, the most important case, no
breakpoints, is just as fast.