[0/13] map->vector: helper function
This patch series replaces a whole lot of std::map's with std::vector's
(sorted or not). The inspiration for these changes came from watching the
following lectures by Alexander Stepanov (creator of STL):
   https://isocpp.org/blog/2014/04/programming-conversations
_Very_ loosely paraphrasing from his lectures (watch the videos for more
context information):

Alexander Stepanov:
  "... I almost regret having included set's and maps in the STL (or having
  them made so easy to use). You can create very complex data structures (map
  of maps) in a single line of C++ code. But many programmers don't understand
  the (performance) cost of these structures. In the majority of the cases a
  vector is a far better solution. No sane C or Fortran programmer would ever
  dream of using a tree based structure when an array suffices..."

With this in mind I went over our code, and indeed many maps could be replaced
by vectors. And it's indeed unfortunate that sometimes the (worse) code using
maps is simpler to write in C++ than the (better) code using vectors. Though
with c++11 this difference has become significantly smaller.

This patch doesn't change any map->vector yet. It only adds a helper function
to make the difference in code complexity even smaller.

--

I doubt any of the changes in this patch series has a big or even a measurable
influence on performance (in the cases where it really mattered we already used
to appropriate structures). Nevertheless it was a good learning experience. We
now also execute approximately 800(!) less heap allocations during startup, so
maybe on small devices it has a small positive impact on startup speed or on
memory consumption.