Fixed crash (only) in debug build
Short summary: self-copy-assignment in c++ is OK (and should have no effect),
but self-move-assignment is undefined. Or in other words:
	a = a;            // ok, no effect
	a = std::move(a); // undefined

This triggered in patch 01b0aa421ea. We have an unsorted vector and want to
erase one element from it. Instead of calling vector::erase() we swap() the to
be removed element with the last element and then do vector::pop_back(). In
this case the vector contains move-only elements, so swap() (must) use
move-assignments. But in case the to be removed element was already in the
last position, this is a self-move-assignment.

Easily fixed by first checking the position before swapping.