Always init video system before getting window handle
This check was performed in DirectXSoundDriver. However, the video
system must always be initialized before the window handle is
requested. So I moved the check into getSDLWindowHandle().

The only other calls to getSDLWindowHandle() are in SDLVideoSystem,
so the video system being initialized is implied there. But it is
safer not not rely on that. Plus this removes the SDL dependency
from DirectXSoundDriver.