Removed libao sound driver
It was a nice experiment, but it never grew beyond that. It didn't
perform better on any platform than the other drivers. Also openMSX
was never adapted to work with an output system that doesn't drop
audio when fed more than it can play, therefore the libao driver would
keep openMSX at 100% speed when throttle was disabled.

Since we are planning big changes in the months ahead, having one less
dependency to worry about will be helpful.