Remove build support for Expat XML parser
This was intended as a lighter alternative to libxml2. However,
integration was never finished. Since we will probably drop libxml2
in the future as well since we barely use it anymore now that there
is an XML parser in src/utils/rapidsax.hh|cc, the chance of Expat
support getting finished is neglible.