Removed build support for SQLite
This was intended as an alternative for parsing the XML-based software
DB at startup. It was never finished though. Instead, we got a faster
XML parser.

We might use SQLite in the future, but then this code can simply be
resurrected from Git. It is also possible we'll be using SQLite through
Qt instead of directly.