lazily create RawFrames in PostProcessor
The previous patch extended PostProcessor with a deflicker filter. This
required to store the last 4 MSX frames. Before we only needed 2 (for
deinterlace). At 32bpp one MSX frame takes 600kB. So two extra frames means
1.2MB extra memory. When gfx9000 is in use the extra memory requirement is even
2.4MB (because v9990 has higher maximum x-resolution).

The change in the previous patch also constantly kept the last 4 frames
up-to-date, even if the deflicker filter wasn't actually used. This means that
a buffer is only being reused every 4+1 frames. This makes it much likely that
it's still in the CPU cache than when the buffer is reused every 2+1 frames
(before previous patch). And when the deinterlace effect is not in use we can
even improve upon the old situation by reusing the memory every 1+1 frames.

So that's what this patch does. It recycles RawFrame objects as soon as possible
(the keep them warmer in the cache) and it lazily allocates them (to safe memory
in the very likely case that the buffers aren't needed).